feat: add default directives
