+++ packages/amplify-codegen/src/commands/models.js
@@ -3,6 +3,7 @@ const fs = require('fs-extra');
 const globby = require('globby');
 const { FeatureFlags, pathManager } = require('@aws-amplify/amplify-cli-core');
 const { generateModels: generateModelsHelper } = require('@aws-amplify/graphql-generator');
+const { DefaultDirectives } = require('@aws-amplify/graphql-directives');
 const { validateAmplifyFlutterMinSupportedVersion } = require('../utils/validateAmplifyFlutterMinSupportedVersion');
 const defaultDirectiveDefinitions = require('../utils/defaultDirectiveDefinitions');
 const getProjectRoot = require('../utils/getProjectRoot');
@@ -131,7 +132,7 @@ const getDirectives = async (context, apiResourcePath) => {
       resourceDir: apiResourcePath,
     });
   } catch {
-    return defaultDirectiveDefinitions;
+    return DefaultDirectives.map(directive => directive.definition).join('\n');
   }
 };
