+++ packages/appsync-modelgen-plugin/src/__tests__/visitors/appsync-typescript-visitor.test.ts
@@ -1,8 +1,11 @@
 import { buildSchema, GraphQLSchema, parse, visit } from 'graphql';
 import { validateTs } from '@graphql-codegen/testing';
+import { DefaultDirectives } from '@aws-amplify/graphql-directives';
 import { TYPESCRIPT_SCALAR_MAP } from '../../scalars';
-import { directives, scalars } from '../../scalars/supported-directives';
+import { scalars } from '../../scalars/supported-directives';
 import { AppSyncModelTypeScriptVisitor } from '../../visitors/appsync-typescript-visitor';
+
+const directives = DefaultDirectives.map(directive => directive.definition).join('\n');
 
 const buildSchemaWithDirectives = (schema: String): GraphQLSchema => {
   return buildSchema([schema, directives, scalars].join('\n'));
