+++ packages/appsync-modelgen-plugin/src/__tests__/visitors/gqlv2-regression-tests/appsync-swift-visitor.test.ts
@@ -1,8 +1,11 @@
 import { buildSchema, GraphQLSchema, parse, visit } from 'graphql';
-import { directives, scalars } from '../../../scalars/supported-directives';
+import { DefaultDirectives } from '@aws-amplify/graphql-directives';
+import { scalars } from '../../../scalars/supported-directives';
 import { SWIFT_SCALAR_MAP } from '../../../scalars';
 import { AppSyncSwiftVisitor } from '../../../visitors/appsync-swift-visitor';
 import { CodeGenGenerateEnum } from '../../../visitors/appsync-visitor';
+
+const directives = DefaultDirectives.map(directive => directive.definition).join('\n');
 
 const buildSchemaWithDirectives = (schema: String): GraphQLSchema => {
   return buildSchema([schema, directives, scalars].join('\n'));
