+++ packages/graphql-generator/API.md
@@ -20,7 +20,7 @@ export function generateModels(options: GenerateModelsOptions): Promise<Generate
 export type GenerateModelsOptions = {
     schema: string;
     target: ModelsTarget;
-    directives: string;
+    directives?: string;
     isDataStoreEnabled?: boolean;
     generateIndexRules?: boolean;
     emitAuthProvider?: boolean;
