+++ packages/graphql-generator/src/models.ts
@@ -1,15 +1,18 @@
 import * as path from 'path';
 import { parse } from 'graphql';
 import * as appSyncDataStoreCodeGen from '@aws-amplify/appsync-modelgen-plugin';
+import { DefaultDirectives } from '@aws-amplify/graphql-directives';
 import { codegen } from '@graphql-codegen/core';
 import { ModelsTarget, GenerateModelsOptions, GeneratedOutput } from './typescript';
 const { version: packageVersion } = require('../package.json');
+
+const directiveDefinitions = DefaultDirectives.map(directive => directive.definition).join('\n');
 
 export async function generateModels(options: GenerateModelsOptions): Promise<GeneratedOutput> {
   const {
     schema,
     target,
-    directives,
+    directives = directiveDefinitions,
     isDataStoreEnabled,
 
     // feature flags
